+++ src/Stryker.Core/Stryker.Core.UnitTest/Initialisation/InitialisationProcessTests.cs
@@ -44,7 +44,8 @@ namespace Stryker.Core.UnitTest.Initialisation
                     ProjectUnderTestAnalyzerResult = TestHelper.SetupProjectAnalyzerResult(
                         references: new string[0]).Object,
                     TestProjectAnalyzerResults = new List<IAnalyzerResult> { TestHelper.SetupProjectAnalyzerResult(
-                        projectFilePath: "C://Example/Dir/ProjectFolder").Object
+                        projectFilePath: "C://Example/Dir/ProjectFolder",
+                        targetFramework: "netcoreapp2.1").Object
                     },
                     ProjectContents = folder
                 });
@@ -90,7 +91,8 @@ namespace Stryker.Core.UnitTest.Initialisation
                     ProjectUnderTestAnalyzerResult = TestHelper.SetupProjectAnalyzerResult(
                         references: new string[0]).Object,
                     TestProjectAnalyzerResults = new List<IAnalyzerResult> { TestHelper.SetupProjectAnalyzerResult(
-                        projectFilePath: "C://Example/Dir/ProjectFolder").Object
+                        projectFilePath: "C://Example/Dir/ProjectFolder",
+                        targetFramework: "netcoreapp2.1").Object
                     },
                     ProjectContents = folder
                 });