+++ src/Stryker.Core/Stryker.Core.UnitTest/Initialisation/InputFileResolverTests.cs
@@ -9,6 +9,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Moq;
+using NuGet.Frameworks;
 using Shouldly;
 using Stryker.Core.Exceptions;
 using Stryker.Core.Initialisation;
@@ -16,6 +17,7 @@ using Stryker.Core.Initialisation.Buildalyzer;
 using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
 using Xunit;
+using static NuGet.Frameworks.FrameworkConstants;
 
 namespace Stryker.Core.UnitTest.Initialisation
 {
@@ -68,30 +70,43 @@ namespace Stryker.Core.UnitTest.Initialisation
         }
 
         [Theory]
-        [InlineData("netcoreapp2.1", Framework.DotNet, 2, 1, null)]
-        [InlineData("netstandard1.6", Framework.DotNetStandard, 1, 6, null)]
-        [InlineData("mono4.6", Framework.Unknown, 4, 6, null)]
-        [InlineData("net4.5", Framework.DotNetClassic, 4, 5, null)]
-        [InlineData("net4.5.1", Framework.DotNetClassic, 4, 5, 1)]
-        [InlineData("net45", Framework.DotNetClassic, 4, 5, null)]
-        [InlineData("net452", Framework.DotNetClassic, 4, 5, 2)]
-        [InlineData("net5.0", Framework.DotNet, 5, 0, null)]
-        [InlineData("net5.0-windows", Framework.DotNet, 5, 0, null)]
-        [InlineData("net5", Framework.DotNet, 5, 0, null)]
-        public void ProjectAnalyzerShouldDecodeFramework(string version, Framework framework, int major, int minor, int? patch)
+        [InlineData("netcoreapp2.1", FrameworkIdentifiers.NetCoreApp, "", 2, 1, 0, 0)]
+        [InlineData("netstandard1.6", FrameworkIdentifiers.NetStandard, "", 1, 6, 0, 0)]
+        [InlineData("net4.5", FrameworkIdentifiers.Net, "", 4, 5, 0, 0)]
+        [InlineData("net4.5.1", FrameworkIdentifiers.Net, "", 4, 5, 1, 0)]
+        [InlineData("net45", FrameworkIdentifiers.Net, "", 4, 5, 0, 0)]
+        [InlineData("net452", FrameworkIdentifiers.Net, "", 4, 5, 2, 0)]
+        [InlineData("net5.0", FrameworkIdentifiers.NetCoreApp, "", 5, 0, 0, 0)]
+        [InlineData("net5.0-windows", FrameworkIdentifiers.NetCoreApp, "windows", 5, 0, 0, 0)]
+        [InlineData("net5", FrameworkIdentifiers.NetCoreApp, "", 5, 0, 0, 0)]
+        public void ProjectAnalyzerShouldDecodeFramework(string tfm, string framework, string platform, int major, int minor, int build, int revision)
+        {
+            var version = new Version(major, minor, build, revision);
+            var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
+                projectReferences: new List<string> { projectUnderTestPath },
+                targetFramework: tfm).Object;
+
+            var nuGetFramework = analyzerResult.GetNuGetFramework();
+            nuGetFramework.Framework.ShouldBe(framework);
+            nuGetFramework.Platform.ShouldBe(platform);
+            nuGetFramework.Version.ShouldBe(version);
+            nuGetFramework.ShouldBe(new NuGetFramework(framework, version, platform, EmptyVersion));
+        }
+
+        [Theory]
+        [InlineData("")]
+        [InlineData("nxt")]
+        [InlineData("mono4.6")]
+        [InlineData("netcoreapp1.2.3.4.5")]
+        public void ProjectAnalyzerShouldRaiseExceptionsForIllFormedFramework(string tfm)
         {
             var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
                 projectReferences: new List<string> { projectUnderTestPath },
-                targetFramework: version).Object;
+                targetFramework: tfm).Object;
 
-            if (patch.HasValue)
-            {
-                analyzerResult.GetTargetFrameworkAndVersion().ShouldBe((framework, new Version(major, minor, patch.Value)));
-            }
-            else
-            {
-                analyzerResult.GetTargetFrameworkAndVersion().ShouldBe((framework, new Version(major, minor)));
-            }
+            Action lambda = () => analyzerResult.GetNuGetFramework();
+
+            lambda.ShouldThrow(typeof(InputException));
         }
 
         [Fact]
@@ -371,9 +386,9 @@ using System.Reflection;
         [Fact]
         public void InitializeShouldNotResolveImportedPropsFile()
         {
-            string sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
+            var sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
 
-            string projectFile = @"
+            var projectFile = @"
 <Project Sdk=""Microsoft.NET.Sdk"">
     <PropertyGroup>
         <TargetFramework>netcoreapp2.0</TargetFramework>
@@ -383,7 +398,7 @@ using System.Reflection;
     <ItemGroup>
     </ItemGroup>
 
-     <Import Project=""../NonSharedProject/Example.props"" Label=""Shared"" />
+     <Import Project=""..\NonSharedProject\Example.props"" Label=""Shared"" />
 
 </Project>";
 
@@ -423,9 +438,9 @@ using System.Reflection;
         [Fact]
         public void InitializeShouldResolveMultipleImportedProjects()
         {
-            string sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
+            var sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
 
-            string sharedItems = @"<?xml version=""1.0"" encoding=""utf-8""?>
+            var sharedItems = @"<?xml version=""1.0"" encoding=""utf-8""?>
                 <Project xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
                 <PropertyGroup>
                 <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
@@ -436,7 +451,7 @@ using System.Reflection;
                 <Compile Include=""$(MSBuildThisFileDirectory)shared.cs"" />
                 </ItemGroup>
                 </Project>";
-            string sharedItems2 = @"<?xml version=""1.0"" encoding=""utf-8""?>
+            var sharedItems2 = @"<?xml version=""1.0"" encoding=""utf-8""?>
                 <Project xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
                 <PropertyGroup>
                 <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
@@ -447,7 +462,7 @@ using System.Reflection;
                 <Compile Include=""$(MSBuildThisFileDirectory)shared.cs"" />
                 </ItemGroup>
                 </Project>";
-            string projectFile = @"
+            var projectFile = @"
 <Project Sdk=""Microsoft.NET.Sdk"">
     <PropertyGroup>
         <TargetFramework>netcoreapp2.0</TargetFramework>
@@ -500,9 +515,9 @@ using System.Reflection;
         [Fact]
         public void InitializeShouldThrowOnMissingSharedProject()
         {
-            string sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
+            var sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
 
-            string projectFile = @"
+            var projectFile = @"
 <Project Sdk=""Microsoft.NET.Sdk"">
     <PropertyGroup>
         <TargetFramework>netcoreapp2.0</TargetFramework>
@@ -512,10 +527,10 @@ using System.Reflection;
     <ItemGroup>
     </ItemGroup>
 
-     <Import Project=""../SharedProject/Example.projitems"" Label=""Shared"" />
+     <Import Project=""..\SharedProject\Example.projitems"" Label=""Shared"" />
 
     <ItemGroup>
-        <ProjectReference Include=""../ExampleProject/ExampleProject.csproj"" />
+        <ProjectReference Include=""..\ExampleProject\ExampleProject.csproj"" />
     </ItemGroup>
 </Project>";
 
@@ -553,11 +568,11 @@ using System.Reflection;
         [Fact]
         public void InitializeShouldResolvePropertiesInSharedProjectImports()
         {
-            string sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
+            var sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
 
-            string sharedFile = "<Project />";
+            var sharedFile = "<Project />";
 
-            string projectFile = @"
+            var projectFile = @"
 <Project Sdk=""Microsoft.NET.Sdk"">
     <PropertyGroup>
         <TargetFramework>netcoreapp2.0</TargetFramework>
@@ -571,7 +586,7 @@ using System.Reflection;
     <Import Project=""../$(SharedDir)/Example.projitems"" Label=""Shared"" />
 
     <ItemGroup>
-        <ProjectReference Include=""../ExampleProject/ExampleProject.csproj"" />
+        <ProjectReference Include=""..\ExampleProject\ExampleProject.csproj"" />
     </ItemGroup>
 </Project>";
 
@@ -616,11 +631,11 @@ using System.Reflection;
         [Fact]
         public void InitializeShouldThrowIfImportPropertyCannotBeResolved()
         {
-            string sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
+            var sourceFile = File.ReadAllText(_currentDirectory + "/TestResources/ExampleSourceFile.cs");
 
-            string sharedFile = "<Project />";
+            var sharedFile = "<Project />";
 
-            string projectFile = @"
+            var projectFile = @"
 <Project Sdk=""Microsoft.NET.Sdk"">
     <PropertyGroup>
         <TargetFramework>netcoreapp2.0</TargetFramework>
@@ -633,7 +648,7 @@ using System.Reflection;
     <Import Project=""../$(SharedDir)/Example.projitems"" Label=""Shared"" />
 
     <ItemGroup>
-        <ProjectReference Include=""../ExampleProject/ExampleProject.csproj"" />
+        <ProjectReference Include=""..\ExampleProject\ExampleProject.csproj"" />
     </ItemGroup>
 </Project>";
 
@@ -754,7 +769,7 @@ Please specify a test project name filter that results in one project.
             var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
             {
                 { Path.Combine(_filesystemRoot, "ExampleProject", "ExampleProject.csproj"), new MockFileData(defaultTestProjectFileContents)},
-                { Path.Combine(_filesystemRoot, "ExampleProject\\ExampleProject2", "ExampleProject2.csproj"), new MockFileData(defaultTestProjectFileContents)},
+                { Path.Combine(_filesystemRoot, @"ExampleProject\ExampleProject2", "ExampleProject2.csproj"), new MockFileData(defaultTestProjectFileContents)},
                 { Path.Combine(_filesystemRoot, "ExampleProject", "Recursive.cs"), new MockFileData("content")}
             });
 
@@ -991,10 +1006,10 @@ Please specify a test project name filter that results in one project.
             var target = new InputFileResolver();
 
             var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
-                    projectReferences: new List<string> { "../ExampleProject/ExampleProject.csproj" }).Object;
+                projectReferences: new List<string> { @"..\ExampleProject\ExampleProject.csproj" }).Object;
 
             var result = target.FindProjectUnderTest(new List<IAnalyzerResult> { analyzerResult }, null);
-            result.ShouldBe("../ExampleProject/ExampleProject.csproj");
+            result.ShouldBe(@"..\ExampleProject\ExampleProject.csproj");
         }
 
         [Fact]
@@ -1005,27 +1020,25 @@ Please specify a test project name filter that results in one project.
                 new InputFileResolver().FindProjectUnderTest(Enumerable.Empty<IAnalyzerResult>(), null);
             });
 
-            ex.Message.ShouldBe("Project reference issue.");
-            ex.Details.ShouldContain("no project", Case.Insensitive);
+            ex.Message.ShouldContain("no project", Case.Insensitive);
         }
 
         [Fact]
         public void ShouldThrowOnMultipleProjects()
         {
             var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
-                    projectReferences: new List<string> {
-                        "../ExampleProject/ExampleProject.csproj",
-                        "../AnotherProject/AnotherProject.csproj"
-                    }).Object;
+                projectReferences: new List<string> {
+                    @"..\ExampleProject\ExampleProject.csproj",
+                    @"..\AnotherProject\AnotherProject.csproj"
+                }).Object;
 
             var ex = Assert.Throws<InputException>(() =>
             {
                 new InputFileResolver().FindProjectUnderTest(new List<IAnalyzerResult> { analyzerResult }, null);
             });
 
-            ex.Message.ShouldBe("Project reference issue.");
-            ex.Details.ShouldContain("Test project contains more than one project reference. Please set the project option");
-            ex.Details.ShouldContain("Choose one of the following references:");
+            ex.Message.ShouldContain("Test project contains more than one project reference. Please set the project option");
+            ex.Message.ShouldContain("Choose one of the following references:");
         }
 
         [Theory]
@@ -1037,14 +1050,14 @@ Please specify a test project name filter that results in one project.
         public void ShouldMatchFromMultipleProjectByName(string shouldMatch)
         {
             var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
-                    projectReferences: new List<string> {
-                        "../ExampleProject/ExampleProject.csproj",
-                        "../AnotherProject/AnotherProject.csproj"
-                    }).Object;
+                projectReferences: new List<string> {
+                    @"..\ExampleProject\ExampleProject.csproj",
+                    @"..\AnotherProject\AnotherProject.csproj"
+                }).Object;
 
             var result = new InputFileResolver().FindProjectUnderTest(new List<IAnalyzerResult> { analyzerResult }, shouldMatch);
 
-            result.ShouldBe(Path.Combine("..", "ExampleProject", "ExampleProject.csproj"));
+            result.ShouldBe(@"..\ExampleProject\ExampleProject.csproj");
         }
 
         [Theory]
@@ -1055,18 +1068,17 @@ Please specify a test project name filter that results in one project.
         public void ShouldThrowWhenTheNameMatchesMore(string shouldMatchMoreThanOne)
         {
             var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
-                    projectReferences: new List<string> {
-                        "../ExampleProject/ExampleProject.csproj",
-                        "../AnotherProject/AnotherProject.csproj"
-                    }).Object;
+                projectReferences: new List<string> {
+                    @"..\ExampleProject\ExampleProject.csproj",
+                    @"..\AnotherProject\AnotherProject.csproj"
+                }).Object;
 
             var ex = Assert.Throws<InputException>(() =>
             {
                 new InputFileResolver().FindProjectUnderTest(new List<IAnalyzerResult> { analyzerResult }, shouldMatchMoreThanOne);
             });
 
-            ex.Message.ShouldBe("Project reference issue.");
-            ex.Details.ShouldContain("more than one", Case.Insensitive);
+            ex.Message.ShouldContain("more than one", Case.Insensitive);
         }
 
         [Theory]
@@ -1076,18 +1088,32 @@ Please specify a test project name filter that results in one project.
         public void ShouldThrowWhenTheNameMatchesNone(string shouldMatchNone)
         {
             var analyzerResult = TestHelper.SetupProjectAnalyzerResult(
-                    projectReferences: new List<string> {
-                        "../ExampleProject/ExampleProject.csproj",
-                        "../AnotherProject/AnotherProject.csproj"
-                    }).Object;
+                projectReferences: new List<string> {
+                    @"..\ExampleProject\ExampleProject.csproj",
+                    @"..\AnotherProject\AnotherProject.csproj"
+                }).Object;
 
             var ex = Assert.Throws<InputException>(() =>
             {
                 new InputFileResolver().FindProjectUnderTest(new List<IAnalyzerResult> { analyzerResult }, shouldMatchNone);
             });
 
-            ex.Message.ShouldBe("Project reference issue.");
-            ex.Details.ShouldContain("no project", Case.Insensitive);
+            ex.Message.ShouldContain("no project", Case.Insensitive);
+        }
+
+        [Theory]
+        [InlineData("ExampleProject/ExampleProject.csproj")]
+        [InlineData("ExampleProject\\ExampleProject.csproj")]
+        public void ShouldMatchOnBothForwardAndBackwardsSlash(string shouldMatch)
+        {
+            var projectReferences = new List<string> {
+                @"..\ExampleProject\ExampleProject.csproj",
+                @"..\AnotherProject\AnotherProject.csproj"
+            };
+
+            var match = new InputFileResolver().DetermineProjectUnderTestWithNameFilter(shouldMatch, projectReferences);
+
+            match.ShouldBe(@"..\ExampleProject\ExampleProject.csproj");
         }
     }
 }