+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/Progress/ProgressBarReporterTests.cs
@@ -1,4 +1,5 @@
 using Moq;
+using Shouldly;
 using Stryker.Core.Mutants;
 using Stryker.Core.Reporters.Progress;
 using Xunit;
@@ -37,6 +38,19 @@ namespace Stryker.Core.UnitTest.Reporters.Progress
                 It.Is<int>(a => a == 3),
                 It.Is<string>(b => b == "│ Testing mutant 0 / 3 │ K 0 │ S 0 │ T 0 │ NA │")
             ));
+        }
+
+        [Fact]
+        public void ShouldSupportWhenNoMutants()
+        {
+            var progressBarMock = new ProgressBar();
+
+            var progressBarReporter = new ProgressBarReporter(progressBarMock, new FixedClock());
+            // the progress bar was never initialized
+            progressBarMock.Ticks().ShouldBe(-1);
+            progressBarReporter.ReportFinalState();
+            // the progress bar was never initialized
+            progressBarMock.Ticks().ShouldBe(-1);
         }
 
         [Theory]
@@ -185,6 +199,20 @@ namespace Stryker.Core.UnitTest.Reporters.Progress
                     expected = "│ Testing mutant 1 / 100000000 │ K 0 │ S 0 │ T 1 │ ~11d 13h │";
                     break;
             }
+        }
+
+
+        [Fact]
+        public void ProgressBarSmokeCheck()
+        {
+            var progress = new ProgressBar();
+            progress.Start(0, "test");
+            progress.Tick("next");
+            progress.Ticks().ShouldBe(1);
+            progress.Stop();
+
+            progress.Dispose();
+            progress.Ticks().ShouldBe(-1);
         }
 
         private void VerifyProgress(string progressBar, int tested, int total, int percentage, string estimate)