+++ src/Stryker.Core/Stryker.Core/Initialisation/InputFileResolver.cs
@@ -26,7 +26,6 @@ namespace Stryker.Core.Initialisation
     /// </summary>
     public class InputFileResolver : IInputFileResolver
     {
-        private const string ErrorMessage = "Project reference issue.";
         private readonly string[] _foldersToExclude = { "obj", "bin", "node_modules", "StrykerOutput" };
         private readonly IFileSystem _fileSystem;
         private readonly IProjectFileReader _projectFileReader;
@@ -75,7 +74,7 @@ namespace Stryker.Core.Initialisation
             projectInfo.ProjectUnderTestAnalyzerResult = _projectFileReader.AnalyzeProject(projectUnderTest, options.SolutionPath, options.TargetFramework);
 
             //to test Fsharp support you would need to create a FsharpProjectComponentsBuilder
-            IProjectComponent inputFiles = new CsharpProjectComponentsBuilder(projectInfo, options, _foldersToExclude, _logger, _fileSystem).Build();
+            var inputFiles = new CsharpProjectComponentsBuilder(projectInfo, options, _foldersToExclude, _logger, _fileSystem).Build();
             projectInfo.ProjectContents = inputFiles;
 
             ValidateTestProjectsCanBeExecuted(projectInfo);
@@ -134,7 +133,7 @@ namespace Stryker.Core.Initialisation
 
         public string FindProjectUnderTest(IEnumerable<IAnalyzerResult> testProjects, string projectUnderTestNameFilter)
         {
-            IEnumerable<string> projectReferences = FindProjectsReferencedByAllTestProjects(testProjects);
+            var projectReferences = FindProjectsReferencedByAllTestProjects(testProjects);
 
             string projectUnderTestPath;
 
@@ -164,19 +163,22 @@ namespace Stryker.Core.Initialisation
             }
         }
 
-        private string DetermineProjectUnderTestWithNameFilter(string projectUnderTestNameFilter, IEnumerable<string> projectReferences)
+        internal string DetermineProjectUnderTestWithNameFilter(string projectUnderTestNameFilter, IEnumerable<string> projectReferences)
         {
             var stringBuilder = new StringBuilder();
+
             var referenceChoice = BuildReferenceChoice(projectReferences);
 
-            var projectReferencesMatchingNameFilter = projectReferences.Where(x => x.Contains(projectUnderTestNameFilter, StringComparison.OrdinalIgnoreCase));
+            var normalizedProjectUnderTestNameFilter = projectUnderTestNameFilter.Replace("/", "\\");
+            var projectReferencesMatchingNameFilter = projectReferences
+                .Where(x => x.Contains(normalizedProjectUnderTestNameFilter, StringComparison.OrdinalIgnoreCase));
             if (!projectReferencesMatchingNameFilter.Any())
             {
-                stringBuilder.Append("No project reference matched the given project filter");
-                stringBuilder.AppendLine(projectUnderTestNameFilter);
+                stringBuilder.Append("No project reference matched the given project filter ");
+                stringBuilder.Append($"'{projectUnderTestNameFilter}'");
                 stringBuilder.Append(referenceChoice);
 
-                throw new InputException(ErrorMessage, stringBuilder.ToString());
+                throw new InputException(stringBuilder.ToString());
             }
             else if (projectReferencesMatchingNameFilter.Count() > 1)
             {
@@ -185,10 +187,10 @@ namespace Stryker.Core.Initialisation
                 stringBuilder.AppendLine(", please specify the full name of the project reference.");
                 stringBuilder.Append(referenceChoice);
 
-                throw new InputException(ErrorMessage, stringBuilder.ToString());
+                throw new InputException(stringBuilder.ToString());
             }
 
-            return FilePathUtils.NormalizePathSeparators(projectReferencesMatchingNameFilter.Single());
+            return projectReferencesMatchingNameFilter.Single();
         }
 
         private string DetermineProjectUnderTestWithoutNameFilter(IEnumerable<string> projectReferences)
@@ -201,14 +203,14 @@ namespace Stryker.Core.Initialisation
                 stringBuilder.AppendLine("Test project contains more than one project reference. Please set the project option (https://stryker-mutator.io/docs/stryker-net/configuration#project-file-name) to specify which project to mutate.");
                 stringBuilder.Append(referenceChoice);
 
-                throw new InputException(ErrorMessage, stringBuilder.ToString());
+                throw new InputException(stringBuilder.ToString());
             }
 
             if (!projectReferences.Any()) // No references found
             {
                 stringBuilder.AppendLine("No project references found. Please add a project reference to your test project and retry.");
 
-                throw new InputException(ErrorMessage, stringBuilder.ToString());
+                throw new InputException(stringBuilder.ToString());
             }
 
             return projectReferences.Single();
@@ -228,7 +230,7 @@ namespace Stryker.Core.Initialisation
             builder.AppendLine($"Choose one of the following references:");
             builder.AppendLine("");
 
-            foreach (string projectReference in projectReferences)
+            foreach (var projectReference in projectReferences)
             {
                 builder.Append("  ");
                 builder.AppendLine(projectReference);